#145 Add: chatInput, dmInput 공동 컴포넌트 messageInput 추가
